Cleaned up custom-redirects types.ts
- types.ts was an ownership-free dumping ground accidentally committed during the recent service refactor; it gave the two types no clear owner
- moved RedirectConfig into its own redirect-config.ts: it is the shared domain type (parser produces it, validation checks it, the service orchestrates it, adapters persist it) with no single consumer, so it gets a dedicated, named home rather than a peer's file
- moved the RedirectsStore port into redirects-service.ts next to the existing RedirectManagerLike: its sole consumer is RedirectsService, so the consumer owns the contract and adapters implement it (correct ports-and-adapters direction)
- type-only change: every other edit just repoints an import; no runtime behaviour changed
