+++ ghost/core/core/server/services/custom-redirects/redirect-config.ts
@@ -1,0 +1,7 @@
+export interface RedirectConfig {
+    from: string;
+    /** Capture groups from `from` can be referenced as `$1`, `$2`, etc. */
+    to: string;
+    /** `true` → HTTP 301, otherwise HTTP 302. */
+    permanent?: boolean;
+}
