+++ ghost/core/core/server/services/custom-redirects/validation.js
@@ -13,7 +13,7 @@ const messages = {
  * non-empty `from` / `to` strings, with `from` compilable as a RegExp.
  * Throws on the first failure rather than collecting errors.
  *
- * @param {import('./types').RedirectConfig[]} redirects
+ * @param {import('./redirect-config').RedirectConfig[]} redirects
  */
 const validate = (redirects) => {
     if (!_.isArray(redirects)) {
